+++ backend/src/main/java/org/yellowcat/backend/product/material/MaterialService.java
@@ -10,16 +10,21 @@ import org.springframework.data.domain.Sort;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Service;
 import org.yellowcat.backend.common.websocket.EntityMessage;
+import org.yellowcat.backend.product.Product;
+import org.yellowcat.backend.product.ProductRepository;
 import org.yellowcat.backend.product.material.dto.MaterialCreateDto;
 import org.yellowcat.backend.product.material.dto.MaterialRequestDto;
 import org.yellowcat.backend.product.material.dto.MaterialResponse;
 import org.yellowcat.backend.product.material.mapper.MaterialMapper;
+
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class MaterialService {
     MaterialRepository materialRepository;
+    ProductRepository productRepository;
     MaterialMapper materialMapper;
     SimpMessagingTemplate messagingTemplate;
 
@@ -38,6 +43,14 @@ public class MaterialService {
 
     public boolean deleteMaterial(Integer id) {
         if (materialRepository.existsById(id)) {
+            Material material = materialRepository.findById(id).orElse(null);
+            List<Product> products = productRepository.findByMaterialId(id);
+            if (!products.isEmpty() && material != null) {
+                material.setStatus(false);
+                materialRepository.save(material);
+
+                return false;
+            }
             messagingTemplate.convertAndSend("/topic/materials", new EntityMessage("Material deleted: ", getMaterialById(id)));
             materialRepository.deleteById(id);
             return true;
@@ -60,5 +73,18 @@ public class MaterialService {
         materialRepository.save(material);
         messagingTemplate.convertAndSend("/topic/materials", new EntityMessage("Material updated: ", material));
         return materialMapper.toResponse(material);
+    }
+
+    public boolean updateStatus(Integer id) {
+        if (materialRepository.existsById(id)) {
+            Material material = materialRepository.findById(id).orElse(null);
+            if (material != null) {
+                material.setStatus(!material.getStatus());
+                materialRepository.save(material);
+
+                return true;
+            }
+        }
+        return false;
     }
 }
