+++ backend/src/main/java/org/yellowcat/backend/product/targetaudience/TargetAudienceService.java
@@ -10,16 +10,21 @@ import org.springframework.data.domain.Sort;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Service;
 import org.yellowcat.backend.common.websocket.EntityMessage;
+import org.yellowcat.backend.product.Product;
+import org.yellowcat.backend.product.ProductRepository;
 import org.yellowcat.backend.product.targetaudience.dto.TargetAudienceCreateDto;
 import org.yellowcat.backend.product.targetaudience.dto.TargetAudienceRequestDto;
 import org.yellowcat.backend.product.targetaudience.dto.TargetAudienceResponse;
 import org.yellowcat.backend.product.targetaudience.mapper.TargetAudienceMapper;
+
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class TargetAudienceService {
     TargetAudienceRepository targetAudienceRepository;
+    ProductRepository productRepository;
     TargetAudienceMapper targetAudienceMapper;
     SimpMessagingTemplate messagingTemplate;
 
@@ -38,9 +43,17 @@ public class TargetAudienceService {
 
     public boolean deleteTargetAudience(Integer id) {
         if (targetAudienceRepository.existsById(id)) {
-            messagingTemplate.convertAndSend("/topic/target_audiences", new EntityMessage("TargetAudience deleted: ", getTargetAudienceById(id)));
+            TargetAudience targetAudience = targetAudienceRepository.findById(id).orElse(null);
+            List<Product> products = productRepository.findByTargetAudienceId(id);
+            if (!products.isEmpty() && targetAudience != null) {
+                targetAudience.setStatus(false);
+                targetAudienceRepository.save(targetAudience);
+
+                return false;
+            }
+
             targetAudienceRepository.deleteById(id);
-            return true;
+            return false;
         }
         return false;
     }
@@ -60,5 +73,18 @@ public class TargetAudienceService {
         targetAudienceRepository.save(targetAudience);
         messagingTemplate.convertAndSend("/topic/target_audiences", new EntityMessage("TargetAudience updated: ", targetAudience));
         return targetAudienceMapper.toResponse(targetAudience);
+    }
+
+    public boolean updateStatus(Integer id) {
+        if (targetAudienceRepository.existsById(id)) {
+            TargetAudience targetAudience = targetAudienceRepository.findById(id).orElse(null);
+            if (targetAudience != null) {
+                targetAudience.setStatus(!targetAudience.getStatus());
+                targetAudienceRepository.save(targetAudience);
+
+                return true;
+            }
+        }
+        return false;
     }
 }
